+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.0.4 (Mar 25, 2020)
+
+* Streamlined mod installation process a bit.
+* Added an erratum for **Stagger** so it's a debuff.
+* Errata'd **Vendetta** so that it doesn't crash when refered to in a card during battle.
+* Buffed **Bloodletting**: Deal additional 3~5 damage and change **Wide Bloodletting** to **Crippling Bloodletting**.
+* Added 2 new battle cards: **Bullseye** and **Provoking Kick**.
+* Added 2 new negotiation cards: **Clairvoyance** and **Surprise information**.
+* Changed the series for **Bloodletting** because Rook can't really use it.
+
 ## 0.0.3 (Mar 23, 2020)
 
 * Added 3 new battle cards: **Bloodletting**, **Bodyguard**, and **Critical Strike**.